+++ packages/log/ts/standardLog.ctx.ts
@@ -1,0 +1,8 @@
+import { record } from 'type-plus'
+import type { LogStore } from './logStore.js'
+import type { LogMethodNames, StandardLogInstance } from './types.js'
+
+export const ctx: {
+  configured?: boolean,
+  gsl?: { store: LogStore, standardLog: StandardLogInstance<LogMethodNames> }
+} = record()