+++ examples/solid/CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ### Patch Changes
 
+- Updated dependencies [bf28064]
+  - standard-log@11.4.1
+  - standard-log-color@11.4.1
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [278afbf]
   - standard-log@11.4.0
   - standard-log-color@11.4.0